+++ database/migrations/2023_11_28_133637_create_users_categories_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('users_categories', function (Blueprint $table) {
-            $table->integer('user_id');
-            $table->integer('category_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('category_id');
         });
     }
